+++ crates/ruma-client-api/src/profile/delete_profile_field.rs
@@ -23,7 +23,7 @@ pub mod v3 {
         authentication: AccessToken,
         history: {
             unstable("uk.tcpip.msc4133") => "/_matrix/client/unstable/uk.tcpip.msc4133/profile/{user_id}/{field}",
-            1.16 => "/_matrix/client/v3/profile/{user_id}/{field}",
+            1.16 | stable("uk.tcpip.msc4133.stable") => "/_matrix/client/v3/profile/{user_id}/{field}",
         }
     }